Move certificate service to separate service part
